pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "obzai"
version = "0.1.1"
description = "Explainable AI, Model Monitoring, and Outlier Detection Tools for Computer Vision Systems"
keywords = [
            "Explainable AI",
            "AI Monitoring",
            "Model Interpretability",
            "Computer Vision",
            "Feature Importance",
            "Explainability",
            "Anomaly Detection",
            "Data Drift",
            "Outlier",
            "Observability",
            "Attention Maps",
            "Saliency Maps"
        ]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.9"
license = { text = "GPL-3.0" }
authors = [
    {name = "Obz AI", email = "hi@obz.ai"}
]
maintainers = [
    {name = "Jakub Binda", email = "binkuba@gmail.com"},
    {name = "Neo Christopher Chung", email = "nchchung@gmail.com"},
]
dependencies = [
    "torch >= 2.1.0",
    "torchvision >= 0.16",
    "transformers[torch] >= 4.46.1",
    "accelerate>=0.26.0",
    "scipy >= 1.3.1",
    "scikit-learn >= 1.6",
    "scikit-image > 0.23.1",
    "numpy >= 1.23",
    "umap-learn >= 0.5.5",
    "pandas",
    "matplotlib",
    "cmasher",
    "tqdm",
    "requests",
    "pillow",
    "captum >= 0.7.0",
    "ipywidgets",
]
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering",
]

[project.urls]
Homepage = "https://obz.ai"
Source = "https://github.com/obzai/obzai"

[project.optional-dependencies]
dev = []

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
